# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TEE Attestation Service API",
    "version": "0.1.0",
    "description": "A service for TEE (Trusted Execution Environment) attestation and key management",
    "contact": {
      "name": "HPE Security Lab",
      "url": "https://github.com/TEE-Attestation/tas",
      "email": "isaac.matthews@hpe.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key for authentication"
      },
      "ManagementApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MANAGEMENT-API-KEY",
        "description": "Management API key for policy administration"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "NonceResponse": {
        "type": "object",
        "properties": {
          "nonce": {
            "type": "string",
            "description": "Generated nonce for attestation",
            "example": "a1b2c3d4e5f6789abcdef0123456789abcdef0123456789abcdef0123456789"
          }
        },
        "required": [
          "nonce"
        ]
      },
      "SecretRequest": {
        "type": "object",
        "properties": {
          "tee-type": {
            "type": "string",
            "enum": [
              "amd-sev-snp",
              "intel-tdx"
            ],
            "description": "Type of TEE"
          },
          "nonce": {
            "type": "string",
            "description": "Previously obtained nonce"
          },
          "tee-evidence": {
            "type": "string",
            "description": "TEE attestation evidence (base64 encoded)"
          },
          "key-id": {
            "type": "string",
            "description": "Identifier of the key to retrieve"
          },
          "wrapping-key": {
            "type": "string",
            "description": "Base64-encoded client public key for wrapping"
          }
        },
        "required": [
          "tee-type",
          "nonce",
          "tee-evidence",
          "key-id",
          "wrapping-key"
        ]
      },
      "SecretResponse": {
        "type": "object",
        "properties": {
          "secret_key": {
            "type": "string",
            "description": "Retrieved and wrapped secret key"
          }
        },
        "required": [
          "secret_key"
        ]
      },
...
    },
...
  },
  "paths": {
    "/kb/v0/get_nonce": {
      "get": {
        "tags": [
          "Key Broker"
        ],
        "summary": "Generate nonce for attestation",
        "description": "Generates a cryptographic nonce for use in TEE attestation",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Nonce generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/kb/v0/get_secret": {
      "post": {
        "tags": [
          "Key Broker"
        ],
        "summary": "Get secret key after TEE verification",
        "description": "Validates TEE evidence and returns the requested secret key",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Secret retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid nonce or API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}